Moved the network and remote universe implementations into their own files
